+++ test/runner.py
@@ -92,7 +92,6 @@ misc_test_modes = [
   'wasmfs',
   'wasm64',
   'wasm64l',
-  'bigint',
 ]